backend/requirements.txt
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
psycopg[binary]==3.3.2
python-dotenv==1.0.1
Pillow==11.0.0
